legacy/alphaflow/Version 5 (Advanced ML & Multi-Asset)/requirements.txt
# Core Dependencies (Required)
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
websockets>=11.0
scikit-learn>=1.3.0
scipy>=1.11.0
pytz>=2023.3

# Trading APIs
alpaca-trade-api>=3.0.0

# Technical Analysis
TA-Lib>=0.4.28

# NEW V5: Deep Learning
torch>=2.0.0
transformers>=4.30.0

# NEW V5: Sentiment Analysis
nltk>=3.8.0
textblob>=0.17.1
newsapi-python>=0.2.7
tweepy>=4.14.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0
